src/backend/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml --no-deps
aiohttp==3.11.12
    # via agentic-voice-assistant (pyproject.toml)
azure-search-documents==11.6.0b4
    # via agentic-voice-assistant (pyproject.toml)
azure-storage-blob==12.23.1
    # via agentic-voice-assistant (pyproject.toml)
cffi==1.17.1
    # via agentic-voice-assistant (pyproject.toml)
chainlit==2.0.6
    # via agentic-voice-assistant (pyproject.toml)
cryptography==44.0.0
    # via agentic-voice-assistant (pyproject.toml)
azure-identity==1.19.0
    # via agentic-voice-assistant (pyproject.toml)
invoke==2.2.0
    # via agentic-voice-assistant (pyproject.toml)
numpy
    # via agentic-voice-assistant (pyproject.toml)
python-dotenv==1.0.1
    # via agentic-voice-assistant (pyproject.toml)
rich==13.9.4
    # via agentic-voice-assistant (pyproject.toml)
streamlit
    # via agentic-voice-assistant (pyproject.toml)
websockets==14.0
    # via agentic-voice-assistant (pyproject.toml)
azure-monitor-opentelemetry==1.0.0b9
    # Added for configure_azure_monitor in Admin.py
tempfile
    # Added for creating temporary directories in 01_Ingest_Data.py